backend/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
requests==2.31.0
websocket-client==1.6.0
pandas==2.0.0
numpy==1.24.0
scikit-learn==1.3.0
tensorflow==2.13.0
python-dotenv==1.0.0
PyYAML==6.0
schedule==1.2.0